Fix browsersync setup within Webpack
Browsersync wasn't creating a correct instance of the app as
the first argument to the plugin was wrong.

This fixes the call to the webpack browsersync plugin so that
it will create the correct URL to auto-refresh changes to the app.

It also specifically sets the port number so that it will always
match what is in the README.