Restructure controller loading in feature apps
This change restructures the way controllers are defined and loaded
within a feature folder app.

It moves to a consistent name for importing controllers into
the index file that manages the router.

When an app increases in size the `controllers.js` file can be
replaced with a `controllres` folder and an `index.js` loader file
without the need to update the require statement that loads the
controllers into the app.

This allows us to be more scalable with the app folder structure
going forward.