Create initial app header variation
This header component uses a the majority of the GOV.UK Frontend
header, including yaml config and tests, but allows the `GOV.UK`
text to be replaced and replaces the crown with the government
crest so that the header can be used for internal services.

This header will likely evolve and hopefuly progress to a higher
Design System for wider use as we discover more use cases.