Support component clashes with GOV.UK Design System
Currently if a local component is created that contains the
same name as one in the GOV.UK Design System it can't be loaded
in as nunjucks looks for the component in the first directory
found with that name in load paths.

To overcome this the component folder can be prefixed with an
underscore to allow it to be treated differently.

This updates the test helpers to accommodate for components that
contain an underscore prefix