Use webpack copy plugin for images
Images were being included via the main JavaScript file which was
causing them to be loaded in the JavaScript when they are actually
only used within the HTML and not wihtin any of the JavaScript.

This switches that method for the webpack copy plugin for copying
the source files to the build directory.