Dynamically load express routers from sub app folders
This change introduces dynamic loading of sub-app routers rather
than needing to manually load them in the router.

This change also introduces support for express mountpaths to
allow apps to be mounted from a particular url path if one is
defined.

This also fixes an issue with middleware defined in sub-apps being
globally loaded and should scope them to only their particular
router.