Display personal details using summary list
This uses the summary list component from the GOV.UK Design System
to display the personal details for the person attached to a move.

This adds the Summary List component to the application and
creates a new presenter to format the person object in a way
the Summary List component.