Update environment variables
This change adds support for using a `REDIS_URL` environment variable
which is the favoured way to connect to Redis in the cloud hosting
solutions this app will use.

It also documents all the environment variables in a more
consistent way and tries to move configuration into the config
module and out of the server.