Store the answers to an assessment step
This change adds the ability to store the asessment answers
on a person object during the form creation session.

Once a step has been successfully saved it will map the values
from the current step and store them as an object on the assessment
property and then also flatten this to the format expected by the
Person service.