Add option to not diplay field in template
This adds support for an optional parameter that will prevent
the field from being displayed in the standard loop of fields.

We will need to be able to add fields to the step but only show
them conditionally. This change allows a parameter of `isConditional`
to be set to `true` which will prevent the field appearing twice.