Configure assessment steps with conditional fields
This creates a new field for each conditional value within an
assessment step.

The field needs to be added to the `fields` on the step so that it
knows to process it.

This method of implementing the commments will involve a level of
configuration on the frontend each time a new question is added
but given the type of conditional content could different I think
this feels safer at this stage.