Use custom error handler for API client
This change overrides the error handler provided by the devour
library.

Previously it would attempt to format error responses into an
object with each field as the key.

Unfortunately it didn't provide any further information with the
formatting, like status codes so other parts of the app wouldn't
know how to handle a validation error (`422`) different to a
genuine server error returned by the API.

It also wasn't being handled correctly by the app error handler as
no stack trace or status code was being passed through.

For now the errors are retained in the same format. This may be changed
in a future change that uses them better to provide form validation
feedback.