Refactor location of form field helpers
The methods in this change were located within the reference data
service but they don't feel related to the service. The function of
them is more related to transforming data between controllers and
view layers.

This change creates a new `helpers` area where these kind of functions
can be stored.