Refactor when assessment conditionals are rendered
Previously the conditional content was rendered into HTML during
the configure step which happens at the start of the lifecycle.

This meant that any steps in between to populate values, or in the
future, error messages were not part of the field object when it
was converted into HTML.

This change moves the rendering of the conditional content into
the final render method of the assessment controller.

This means the fields can be manipulated during the lifecycle as
other fields are before it is rendered into HTML.