Ensure mapped object remains the same in helpers
When mapping the field objects in helpers like renderConditionalFields
the item was being changed from an Array to an Object.

This meant it wasn't possibly to chain more maps to the initial call.

This change ensures the mapped item is returned in the same format
so that it can be further manipulated.