Ensure only requested moves are displayed on dashboard
When we extend the app to allow moves statuses to be changed, for
example if a user cancels a move, we don't want those other statuses
to appear on the main dashboard.

This change filters the moves displayed on the dashboard to only
include the status it cares about (`requested`).