Restructure URLs to be more logical
Currently we are serving a list of moves on the route URL within
a dashboard feature folder.

All "moves" related content is then within the moves feature folder.

This change restructures the dashboard so that the moves are served
under `/moves` instead of `/`. This now makes more logical sense
with the rest of this feature.

This change also adds a redirect for the route path to the moves
list for now. As the app grows the route path may be used for other
actions.