Switch i18n translation library
This change replaces the i18n-future library with i18next.

The i18next library has a lot more features, including substituting
values using either strings or objects passed to the translation
method call.

These are needed in certain parts of the app to be able to fully
extract all the content.