Ensure formatTime doesn't return a time for falsey values
Previously when `formatTime()` was given a `null` or `false`
value it was parsing the date as `1970-01-01T00:00:00.000Z`
which would cause it to return a time for falsey values.

This change ensures it won't return a valid time and adds some
more tests to prevent this happening in the future.