Move auth specific middleware into the auth feature
This change moves some of the logic that was only needed during
auth callback handling into the auth feature folder and keeps
the app level middleware in the common folder.

It also does some work to simplify some of tests and remove the
need for the handler to be an async function.