Move getAssetPath helper into Nunjucks globals
Currently this helper is set within a locals middleware. In some
instances when an error is thrown during a middleware before this
middleware the error template doesn't render correctly because
this method doesn't exist.

There is actually no need for it to be in the locals middleware as
it doesn't require access to the `res` or `req` variables that
are made available via middleware.

With the `getAssetPath` method now always being available the error
template won't throw an error.