Prevent error template errors
1) Re-order i18n middleware to be loaded earlier as
translations are used within the error template. This prevents
nunjucks throwing an error when trying to render the error view.

2) Only call getMessages if it actually exists in locals. getMessages
relies on the request object so needs to be loaded lower down after
the session and flash library and the current template call would
throw an error if it hadn't got that far down the chain