Splits moves into separate feature folders
This change refactors the moves feature folder into two folders. One
to list/download the moves and another to view/create/cancel an
individual move.

This split is a precursor to some changes coming around restricting
access to particular actions based on a user's permissions.

The separation also helps to reduce the size and complexity of the
current moves feature folder.