Redirect a user based on permissions
This change decides where to send a user based on their permissions.

If a user has the ability to view all moves they will be redirected
to the moves base URL.

If a user does not have view all permission, but has view by location
permission and a location is set they will be redirect to that location.

If a user is missing both permissions they will get a 404 at the root
path.