Fix auth return URL
Currently with the order in which the files are loaded the
auth logout return URL isn't being populated.

The url helper was being loaded before the config had populated any
values that were loaded into that module and set as defaults.

This change simplifies that to build the base url in the config
file which can be sent to the auth logout as a return URL.