Catch missing authorities property on auth token
When an auth token is returned from the current auth provider (HMPPS SSO)
and that user doesn't have a role assocaited with it, there will be
no `authorities` property on the token.

When the user is missing this property the User class would throw an
error as it expected it to exist.

This fix will set a default value if that property doesn't exist and
return an empty permissions array.