Move permissions redirect to moves feature
This change moves the logic that determines where to redirect a
user based on their permissions to the moves feature folder rather
than the route.

The current logic meant that a user could sign in with no permissions
and would see a `404` at the route so wouldn't necessarily know its
a permission issue.

Moving to the moves feature folder will have a user fallback to the
protectRoute middleware if they have no permissions to do anything
within the moves feature.