Add created date to CSV download
This adds a created at column to the CSV download.

It currently uses `updated_at` from the API but we can assume
this is the created date because we don't currently support any
update functionality on the frontend.

Have added a todo to ensure this is updated when the API supports
`created_at` as well as `updated_at`.