Update namespace separator for i18n
The default namespace separator for the i18n library is currently
`:`.

This has raised an issue with some strings that don't have translations
including a single colon and therefore looking for a particular
namespace first.

When this is not found it removes that first part of the string
and doesn't fallback to the full string.

This caused an issue with NOMIS imported ethnicities which included
a single colon.

This change replaces the single colon with a double with the
assumption that this will less likely occur in fallback strings.