Add page title to form wizard
This change sets the page title for form wizard steps. At the moment
the form wizard falls back to the default page title.

This change adds a new property to the wizard settings to use as
the page title. It also renames the `heading` property for each step
to be more consistent.