Fix pagination alignment
The inline variation of the pagination component wasn't being
displayed correctly on certain browser (Safari) due to the position
the inner elements were being given.

The text element was being `display: inline-block` which caused the
text deocration not to be inherited which concealed the alignment
issue.

This change displays the elements as inline and sets the text decoration
styles on the correct element (the parent anchor).