Ensure API auth call includes timeout
A timeout was added to all regular API client calls but the
separate auth call was missed.

This change adds the setting to the auth middleware to ensure it
receives the same timeout. This is an initial fix to prevent
the frontend from hanging.

In future it might also be better to make use of the axios instance
on the jsonApi object. But that can come as part of a refactor to the
way the API client is currently tested and used within the app.