Refactor API model definitions
This change exports the model objects rather than a function that
defines each model on the API client.

This will allow the model and its attributes to be used in other
parts of the app like services to ensure when formatting they follow
the same property structure as the API model.