Increase page limit on moves request
This change ensures that we return the maximum number of moves
to the dashboard. At the moment there is no pagination so this
will ensure that no moves are not displayed on the dashboard.

There will be a performance issue with this initially but this
has been accepted for this initial pilot where only small number of
moves are being put through a day.