Fix moves permissions with no locations case
If a user has permission to view all moves, regardless of location,
they should be able to view moves even if their user does not currently
have any available locations.

This is currently the case for suppliers. Without this change a
supplier user is only greeted with the no locations message and cannot
view any other parts of the app.