Ensure session is destroyed before redirecting
The session destroy method as asynchronous, especially when using
a session store like Redis.

Currently the user is being redirect before the session is actually
destroyed which has led to some inconsistencies when signing out
and back in again in quick succession.

This change calls the redirect in the destroy callback to ensure
it has been destroyed properly.