Refactor move service unit tests
This refactors the way the move service unit tests work.

Currently they rely on mocking the HTTP requests from the API
client which has lead to some vague tests.

This refactor stubs the API client library so that the tests only
cover this layer.

It also means that the need to stub the auth middleware no longer
needs to take place.