Remove need for config item default
Previously we needed to set a default value for `API.BASE_URL` as tests
were relying on this value.

Those tests have now been refactor so that this default can be removed
safely.

This change also removes the config test which currently relies on
the default value. This will throw up a reduction in test coverage
for the config file but this should be addressed differently rather
than relying on environment variables being set in test environments.

This changes means all unit tests can run in any environment without
requiring environment variables to be set.