Ensure person service transforms the response
The transformation of the person model response was moved to a
shared place so that we can do common things with each Person.

This was added to the Move service where a person exists on a move
but it wasn't added to the Person service.

This change ensure that all responses from the API regarding creating
or updating a person transform the response in the same way.