Ensure a user can only see locations they have access to
This change amends the location middleware to check that the user
can see this location before displaying the view.

Previously if a user knew the UUID of a location that they didn't have
access to they would've still been able to see that location's moves.

Now if a user accesses a UUID that is not found in their list of
permitted locations it will return a `404`.