Add sticky sidebar mobile override styles
When the sidebar is used in small screen view it continues to use the
JavaScript to make the sidebar sticky. This functionality isn't required
on small screen so this work overrides the styles being applied by The
JavaScript.

Of note this is not a great fix but feels like the better option in
comparison to detecting screen size with JavaScript.