Update API environemtn variable configuration
This change amends the way the API is set in the environment variables.

Currently it requires the host to be defined 3 times for `API_BASE_URL`,
`API_HEALTHCHECK_URL` and `API_AUTH_URL`. If the host needs to be
changed, for example switching between environments when testing this
requires changing 3 variables.

This change just neatens that up so that it only needs to be defined once.

It introduces some temporary logic to make sure the change is backwards
compatible with current releases. Once all envionrments have been
updated this logic can be removed.