Add move date to URL path rather than query
This change moves the move date used to get the list of moves to
a URL param from a query string param.

The move date is crucial to displaying all of these pages so it
makes more sense for it to be part of the URL rather than as a part
of the search string.