Only allow user's to see their locations
Currently the root moves URL with date will allow a user to see all
moves in the system.

This change ensures they can view a collection of moves but only for
locations that they have access to.

This change is to allow suppliers to download a list of their moves
for all locations they have access to.