Multi upload file component
Component initially taken from MoJ design system
https://moj-design-system.herokuapp.com/components/multi-file-upload

But totally reworked as the component wasnt production ready and had a
number of things not quite finished with it.

This component is now ready for production and works with and without
JavaScript. It has been written in vanilla JavaScript and works with
keyboard.