delete move if it exists when journey canceled
Because the move has to be created before documents are uploaded to it,
because documents require a moveId to be uploaded to. A user could
cancel the move in the Create a move journey after it has been created
and before documents are uploaded.
This work will delete a move if it exists and it has been canceled.