Bring the parseMultipartForm work into documents
The parse multipart form work is not needed elsewhere in the app and it
is sharing a few things found in the documents controller. For the
moment it felt sensible to bring this into the documents controller
rather than abstracting it away. This may change in the future if we
require multi part work elsewhere in the app