Reset the file input
When a user uploads files via the "choose files" button with JavaScript
enabled the onChange event loads the files by Ajax. But because the
files also remain in the #file input the files will upload again when
the form submits.
This work stops the files from being submitted on clicking of the
"schedule move" button in this situation