Improve waiting for CSV downloads helpers
This work improves the CSV waiting for downloads methods. Basically it
now waits until the glob returns a result. When the glob returns a
result it means the files have downloaded and exist. This simplifies the
work that was there previously and removes the hard wait of 5 seconds
which at some point would more than likely fail.