Ensure current location is persisted on session regeneration
Currently there is a bug where if the session expires on the frontend,
it will attempted to redirect to the auth service which will make a
call to see if the user is still authenticated.

If they are, it will regenerate the session, persisting some values.

However, the `currentLocation` value is not persisted across this
regeneration which means that the check locations middleware then
redirects the user back to the locations mountpath to choose their
location again even though, in their eyes, they were never logged
out.

To fix this, `req.session.currentLocation` is now persisted across
this session rengeneration so that the user remains on the page they
were before the re-authentication.