Move button logic to controller
This change moves the logic to set the correct button text for the
create journey, whether to display continue or schedule move, to
the controller so that this logic doesn't need to live in each step.

It also makes it easier when using the forking via next as the button
text will be dynamically set based on the next step of each current
step.