Remove destroy method from document service
As the delete method on the API doesn't actually delete a document
we don't need to call this method during the delete process in the
create journey.

Instead we can just remove that document from the session so that it
isn't sent to the API to associate to the move.