Only run production build/deploy jobs on Git tags
We use git tags when creating a new release to deploy to production.

So this change will mean the approval job will only run when new
git tags are pushed.

This prevents it running on every commit to master which is currently
returning a 'pending' status for the workflow on CircleCI.