refactor(createmove): Move setting of location to base controller
This moves where we set the values for `from_location`, `from_location_type`
and `can_upload_documents`.

Previously they were done in the move details controller but seeing as
we know them at the start we an move them to the base controller and
ensure they're available from the beginning of the journey.