feat(createmove): Add logic to handle no prison results
When a user searches for someone with a prison number and there are no
results we need to ensure that they can't create a new person in our
system.

This adds logic to not allow a user to progress to the personal
details step.

It also includes some changes to the display logic of the messages
depending on if there are results or not.