refactor: Remove the need for user roles
This refactor removes the only current use of a user's role within
the codebase - sending journey times to analytics.

This logic is also no longer what we use to determine different journeys.

It also includes reworking the location of it to be confined to the
actual create journey and its session so we don't need to manipulate
the session within an analytics library and avoids us neeeding to pass
in the request object.