fix: Ensure date field runs formatter
The date refactor change the formatter to use a string rather than
the function that it needed to. This meant that certain date strings
that were previously supported no longer were supported.

This ensures the formatter is called correctly and uses the shared
date field object with other date fields, like date of birth.

It also includes an update to the e2e tests to catch if this changes
again in the future.