refactor: Move updating of assessment answers to end of journey
This refactor moves the updating of the assessment answers on the
person object to the end of the journey and until the final save.

This is so that when using the person lookup, if the person has any
previous answers they are not overriden until the last step. And
allows us to replay any existing answers back to the view.