fix: Update prison number key
Previously we were using a different filter for the prison number
identifier. Trying to alias `nomis_offender_no` to `prison_number`
on the API caused issues so we're rolling the change back to only
consistently use `prison_number` for now.

This change updates the filter on the frontend to use the new API
filter.