refactor: Move configuration of step logic to steps file
This change moves how the different fields are set within the move
details step.

We were doing a lot of logic in the initial configure method of the
controller. This adjusts some of that and makes use of the middleware
hooks within the form wizard to reduce the complexity of the methods,
also making them easier to test.