fix: Add support to define keys to show for previous assessment
This change allows a step to define what assessment answer keys should
be played back to the user for that step.

Before it relied on showing only fields that were in the step, but in
one case we also wanted to include pregnancy in the list of answers
we show.