refactor: Refactor e2e tests to use page object model
This refactors the e2e tests to use a more shared page object model
approach.

This makes it easier to re-use more actions and if elements change
it means they only needed to be updated in the page objects.

There is still some more to refactor but this is an initial start at
making them easier to re-use for the upcoming journeys.