refactor: Allow images to be controlled by env var
This change reworks how images are retrieved from NOMIS. It introduces
a new environment variable that will allow us to turn the image sync
to NOMIS on or off by an environment variable (`FEATURE_FLAG_IMAGES`).

Once we decide we always want these on again we can revert this commit.