refactor: Mock dependent libraries in existing test
This refactor removes the dependency on the presenter from this test.

Previously it was calling the presenter which meant that if things like
field labels were updated they needed to be updated here.

This test should be more isolated and only test the function within it,
like that it calls the presenter with the correct values and sets
the correct values on the locals object.