fix: [P4-1041] Extend service timeout
Problem
-------

After successfully reauthenticating a user, the user's session is regenerated. Unfortunately this had the side-effect of losing all the info that HMPO form wizard has stored resulting in the user seeing a session timeout message (or for xhr requests, the cryptic 'Server returned code 0')

Even if the HMPO form wizard was not being lost, the reauthentication redirect would have the effect of causing any posted information to be dropped.

Ideally a new token could just be requested using the refresh token rather than redirecting the user, but the HMPPS OAuth server does not support refresh tokens for the type of grant we require.

Fix
---

- Copy properties from previous session to new one
- Stash POST body (if any) in `req.session.originalRequestBody` when redirecting to /connect/$provider route
- Return 401 for request types that cannot be handled
  - when content-type is `multipart/formdata` (as we can't store the uploaded file)
  - when request is an xhr request (since the xhr runs afoul of cross-site concerns)
- Update request's body and method if `req.session.originalRequestBody` is present after successful reauthentication
- Refresh the auth token if the request method is 'GET' and the token is about to expire (see AUTH_EXPIRY_MARGIN env var)

Env vars
--------

Add  `AUTH_EXPIRY_MARGIN` to determine how much time should remain for the auth token to expire before attempting to refresh it. Defaults to 300 seconds (ie 5 minutes)