fix: Prevent field helper mutating properties
This helper method was mutating the original field and assigning
the dependent property. This meant in the case where lots of fields
existed, each field got the value of the last item.

This change creates a copy and re-assigns the field rather than
adding a property to the original field object.