refactor(e2e): Remove use of snake case in e2e tests
We were using a lot of snake case which was causing more mixed cases
to creep in to lots of the e2e test files.

We do still need to use snake case to deal with the API response when
creating a fixture so this has been isolated to the fixture helper
method.