fix(e2e): Handle case where no moves are in the download
This fix ensures that the tests don't fail if the location that is
being downloaded doesn't have any moves.

Previously it was comparing the header line to the first line of the
CSV. Now it will get the length of all the rows and ensure they match.

This will also catch any errors in any rows after the first row.