chore: Update hearings to correct attribute name
This change updates the court hearings attribute to match the new
attribute coming from the API.

When this change was originally merged the API returned hearings as
`hearings` but after some changes to make hearings more explicit
and clear it has been updated to `court_hearings`.

This change updates the references to the old attribute.