refactor: Change permissions for viewing/downloading moves
This refactor applies the new models and navigation to permissions.

Previously permissions where done by `all` or `by_location`. But now
we have different levels of what we mean by "by_location".

We are also trying to introduce a new top level navigation and it makes
sense to build the items available using permissions. So this change
introduces an "outgoing" in place of "by_location" and reduces the ability
to downloads moves to a single permission.