refactor: Update upcoming moves to outgoing moves
This changes the language we use to be more clear about what we are
actually showing the users. Previously we have referred to them as
"upcoming" moves but this would include both outgoing and incoming.

Our current screen only shows outgoing so this change updates the content
to better reflect that.