fix: Use fallback config when creating Redis store
This fix moves the config into the redis store library so that if
it is created without config, it will use the settings from the config
file.

This means the load order is less important when creating the client.

It also means that it can be used elsewhere without needing to pass in
the configuration, for example, without end-to-end tests.