fix: Update consistency of "requested" language
The statuses for a move were recently updated. We used to refer to a
move as "Scheduled" when it was created, but with the introduction of
different states for a supplier this term was changed to "Requested".

Currently we use a mix of the two terms which could be confusing to users.

This change ensures we have consistency in terminology and updates
any existing references to "Scheduled" to be "Requested".