refactor: Extend permissions to determine permitted destinations
This refactor moves some of the existing logic from the steps object
to one of the controllers.

It moves the logic for determining which destinations a user can send
a person to. That logic has had to be based on permissions to some extent
which meant creating various different "steps" and using the Form Wizard
forking logic to handle that.

This caused a problem when trying to introduce the ability to go back
and make a change to the move details when there is a conflict as there
isn't one step to direct a user to.

The existing logic also meant transforming the name of one of the fields
to not break the logic within the process method of the lifecycle.

This change uses permissions to determine which types of move a user can
create and builds the move type options from there. This allows us to link
to one step if making changes and reduces some of the complexity in the
steps and fields configurations.