refactor: Make markup for assessment shared in a presenter
This change refactors how the assessment answers are built. Currently
they are built within the view, in two different places. There is a
change needed to show them in a third place/way so this code is being
refactored into a shared place so that it can be reused for that too.