chore: Add support for getting a person's timetable
This chore adds a new model to the API layer to retrieve a timetable
as well as adding a new method to the Person service to allow a
person's timetable to be retrieved for a given date.

This work is needed to introduce the new step to show possible
conflicts when creating a `prison to court` move.