refactor: Update move detail banners
This updates the header section of the move detail page to use banners
less frequently and only for particular states. It reverts some changes
to introduce the banner for all statuses.

Now it uses the same style as on the dashboard to provide more continuity
across pages.

Particular states which need to draw more attention, like a cancelled
state, now use the temporary notice banner so that they are more obvious
that the move is in a unique state.