fix: Split date range filters so current week is shown correctly
Currently we use the same filter for displaying a range of dates and
displaying a range of dates as a relative week.

This doesn't give us control to decide whether in a certain context
we want to display that range as a relative week.

In this instance, if two dates for a move where the same as a current week
it would show "This week" rather than those dates in the table format.

This splits the filters giving us the control of when we want a date range
to be displayed relatively to "this week".