fix: Format prison transfer relationships
The prison transfer reason is currently a relationship in the API, so
when we are sending that data from something like the formwizard to the
service we need to make sure the value is structured correctly as an
object.

Otherwise the devour serializer will not send it correctly to the API
and will not store the value correctly.