fix: Update how from and to dates are handled in the summary panel
Previously we were combining the from and to date into the one date
row on the right hand move summary panel.

We've decided to split this into separate rows to give a clearer picture
of the situation.

This also fixes a bug where the mock date happened to be tomorrow and
we weren't mocking the date formatters which caused the text to be
different and the tests to start failing.