refactor: Use separate field for prison recall comments
This refactors how we capture the comments associated with a prison
recall move during the form wizard journey.

Previously we used the additional information field. This now also
needs to be re-purposed by another part of the journey so this field
is being given its own name and conditions being set as to whether
it saves its value as additional value.