refactor: Remove need to pass in translate function
For some of the field helpers that require translation we previously
passed in the translate function via `req.t` as that's where the
i18n library added it via express middleware.

The i18n has since been made available as a separate library so we can
now call it directly within a module. This refactor removes the need
to pass it in from the controllers.