fix: Ensure court hearing start time is in the future
Previously we were allowing a court hearing time to be in the past
if a user has set the move the move date to "Today".

This means the court date will not be written to NOMIS.

This change validates that the start time, along with the move date
are in the future.