refactor: Update dashboard to remove time period pagination
The dashboard will start to link to sections that have different
time periods. Therefore we are starting to remove the navigation
that will seem confusing when having different time periods.

The dashboard will start to be more of a starting block to enter other
sections of the service and the panels will show a summary for
pre-defined, common, time periods.