fix(single requests): Ensure filter is only hidden when no total moves
Currently the filter is not displayed if there are no moves, as there
is nothing to filter.

However, the view doesn't no about the total number of moves so if
a user navigates to a tab with no moves, the filter isn't displayed.

This change passes the total number of moves for that day to the view
as well as the moves for the table. Both no have different "empty"
states.