fix: Update end-to-end tests to support new "home" dashboard
The root dashboard was changed to remove the pagination and to act
more like a springboard into other parts of the service. This broke the
existing smoke tests.

This change adapts the smoke tests to cater for a different style of
dashboard and introduces a new page object to deal with them.