fix: Handle how CSRF errors are dealt with
This change improves how CSRF errors are dealt with. At the moment,
they fallback to the default error handler which causes a user to see
a "Page unavailable" page and trigger a sentry error.

This error case is expected and we can provide a better route to handling
it. This change catches the error as part of the form wizard and displays
the form wizard error page. This will also prevent us from seeing this
error in sentry which should be reserved for unexpected errors.

This fixes sentry error #BOOK-A-SECURE-MOVE-FRONTEND-10