chore(moves): Allow the moves service getAll to return an aggregation
This adds support to return an aggregated count for a call to the moves
endpoint with a particular filter.

It reduces the amount of items to load per page to reduce the size of
the response.

This is an MVP solution to support other methods passing this value
through to this method.