chore(single requests): Add new service to deal with single requests
This creates a new service to start handling single requests.

They are becoming a separate entity with a unique way of handling them
within the API which requires abstraction. Previously we were
trying to display them based on just a move status, but we now need
to filter them on other criteria to return the correct results.

This creates an extra layer on top of the moves service to deal
specifically with single requests in different states within the UI.