fix(single requests): Fix single request filtering
This refactor uses the new single requests service to return
data related to single requests.

Previously we were making use of the existing moves service and just
passing a status, but single requests require some extra filters
within the API so a new service was created to create that abstraction.

This change removes the old services and replaces them with this new one.